* Created a valid / success banner to display when a valid tangram is created   or a puzzle is solved, replacing the pop-up Toast in the latter case. * In free-form mode if the player makes a valid tangram but   does not have a save directory set, display a small version   of the checkmark icon in place of the save button. * Changed the piece and outline colors for the .Puzzle and .Hinted themes to   make sure they're not using any transparency that could reveal the outlines. * Doubled the precision of TPoint's coefficients for better accuracy. * Added TPoint.subtract(TPoint) to replace .mirrorX().mirrorY(). * Added TPoint.toImmutable() for places where we   need to ensure a shared result isn't changed. * Added an optional "pointOfRotation" field to TangramPiece'   JSON representation to simplify the representation of tans   placed at angles other than multiples of 45°.  Started   modifying the affected puzzle definitions accordingly. * Modified ImmutableTPoint.rotate(d) to pass off rotations not   in multiples of 45° off to MutableTPoint.rotate(d).  Added   special cases in MutableTPoint.rotate(d) to handle fine   rotations of exactly 15° or 30° by moving all the components   that aren't exact multiples of √2 into the 'a' coefficients,   and for fine rotations that are not multiples of 15° put   the whole result into 'a' and zero out the 'b' coefficients.   Added more test cases to TPointTest to verify the results. * Added a debug line to PlayActivity to show whether and why a   puzzle failed to validate in LogCat if all pieces are present. * BUG FOUND: TPolygon may evaluate different loops   depending on the order in which pieces were added.

+++ app/src/main/java/com/xmission/trevin/android/tangram/ui/PieceTrayItemView.java
@@ -297,15 +297,15 @@ public class PieceTrayItemView extends FrameLayout {
         float minX = Float.MAX_VALUE, minY = Float.MAX_VALUE;
         float maxX = -Float.MAX_VALUE, maxY = -Float.MAX_VALUE;
         for (TPoint v : vertices) {
-            minX = Math.min(minX, v.getX());
-            maxX = Math.max(maxX, v.getX());
-            minY = Math.min(minY, v.getY());
-            maxY = Math.max(maxY, v.getY());
+            minX = (float) Math.min(minX, v.getX());
+            maxX = (float) Math.max(maxX, v.getX());
+            minY = (float) Math.min(minY, v.getY());
+            maxY = (float) Math.max(maxY, v.getY());
         }
         Path path = new Path();
         for (int i = 0; i < vertices.length; i++) {
-            float px = (vertices[i].getX() - minX) * unitScale;
-            float py = (vertices[i].getY() - minY) * unitScale;
+            float px = (float) (vertices[i].getX() - minX) * unitScale;
+            float py = (float) (vertices[i].getY() - minY) * unitScale;
             if (i == 0)
                 path.moveTo(px, py);
             else
